+++ config/routes.rb
@@ -1,6 +1,10 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   root 'welcome#index'
-  # post "/feeds", to: "feeds#feeds"
-  resources :feeds, only: [:create]
+
+  namespace :api do
+    namespace :v1 do
+      resources :feeds, only: [:create]
+    end
+  end
 end
